Ajout du menu pop-up
