+++ desktop/src/civ/game/DesktopLauncher.java
@@ -2,7 +2,7 @@ package civ.game;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import civ.game.civgame;
+import civ.game.CivGame;
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
 public class DesktopLauncher {
@@ -11,6 +11,6 @@ public class DesktopLauncher {
 		config.setForegroundFPS(60);
 		config.setWindowedMode(600, 400);
 		config.setTitle("civgame");
-		new Lwjgl3Application(new civgame(), config);
+		new Lwjgl3Application(new CivGame(), config);
 	}
 }
